Fix #94, move all bitfields into codec
Nothing outside of the codec code should need to know about the bitfields
in the CFDP protocol definition.  Everything else uses logical values.

This moves all DECLARE_FIELD macros into cf_codec.c and removes cf_field.h.
