+++ unit-test/cf_cfdp_s_tests.c
@@ -3,7 +3,6 @@
 #include "cf_test_alt_handler.h"
 #include "cf_cfdp.h"
 #include "cf_app.h"
-#include "cf_field.h"
 #include "cf_events.h"
 
 #include "cf_cfdp_r.h"
